pyproject.toml
[project]
name = "jinja2-tester"
version = "0.1.0"
description = "A web application for testing Jinja2 templates"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask==2.3.3",
    "Jinja2==3.1.2",
    "PyYAML==6.0.1",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jinja2_tester"]
